+++ src/app/Http/Controllers/DashboardController.php
@@ -4,18 +4,22 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Http;
+use Throwable;
 
 class DashboardController extends Controller
 {
+    protected $harvey_domain_protocol;
     protected $harvey_domain;
     protected $harvey_secret;
     protected $timeout;
 
     public function __construct()
     {
+        $this->harvey_domain_protocol = getenv('HARVEY_DOMAIN_PROTOCOL') ?? 'http';
         $this->harvey_domain = getenv('HARVEY_DOMAIN');
         $this->harvey_secret = getenv('HARVEY_SECRET');
-        $this->timeout = 10;
+        $this->timeout = getenv('HARVEY_TIMEOUT') ?? 10;
+        $this->harvey_page_size = getenv('HARVEY_PAGE_SIZE') ?? '20';
     }
 
     /**
@@ -25,25 +29,41 @@ class DashboardController extends Controller
      */
     public function dashboard()
     {
-        $harvey_health_response = Http::withBasicAuth($this->harvey_secret, '')
-            ->timeout($this->timeout)
-            ->get("http://$this->harvey_domain/health");
-        $harvey_status = $harvey_health_response->status();
+        try {
+            $harvey_health_response = Http::withBasicAuth($this->harvey_secret, '')
+                ->timeout($this->timeout)
+                ->get("$this->harvey_domain_protocol://$this->harvey_domain/health");
+            $harvey_status = $harvey_health_response->status();
+        } catch (Throwable $error) {
+            $harvey_status = 500;
+        }
 
-        $projects_response = Http::withBasicAuth($this->harvey_secret, '')
-            ->timeout($this->timeout)
-            ->get("http://$this->harvey_domain/projects");
-        $projects = $projects_response->successful() ? $projects_response->json()['projects'] : [];
+        try {
+            $projects_response = Http::withBasicAuth($this->harvey_secret, '')
+                ->timeout($this->timeout)
+                ->get("$this->harvey_domain_protocol://$this->harvey_domain/projects?page_size=$this->harvey_page_size");
+            $projects = $projects_response->successful() ? $projects_response->json()['projects'] : [];
+        } catch (Throwable $error) {
+            $projects = [];
+        }
 
-        $pipelines_response = Http::withBasicAuth($this->harvey_secret, '')
-            ->timeout($this->timeout)
-            ->get("http://$this->harvey_domain/pipelines");
-        $pipelines = $pipelines_response->successful() ? $pipelines_response->json()['pipelines'] : [];
+        try {
+            $pipelines_response = Http::withBasicAuth($this->harvey_secret, '')
+                ->timeout($this->timeout)
+                ->get("$this->harvey_domain_protocol://$this->harvey_domain/pipelines?page_size=$this->harvey_page_size");
+            $pipelines = $pipelines_response->successful() ? $pipelines_response->json()['pipelines'] : [];
+        } catch (Throwable $error) {
+            $pipelines = [];
+        }
 
-        $locks_response = Http::withBasicAuth($this->harvey_secret, '')
-            ->timeout($this->timeout)
-            ->get("http://$this->harvey_domain/locks");
-        $locks = $locks_response->successful() ? $locks_response->json()['locks'] : [];
+        try {
+            $locks_response = Http::withBasicAuth($this->harvey_secret, '')
+                ->timeout($this->timeout)
+                ->get("$this->harvey_domain_protocol://$this->harvey_domain/locks?page_size=$this->harvey_page_size");
+            $locks = $locks_response->successful() ? $locks_response->json()['locks'] : [];
+        } catch (Throwable $error) {
+            $locks = [];
+        }
 
         return view('/harvey', compact('harvey_status', 'projects', 'pipelines', 'locks'));
     }
@@ -57,10 +77,14 @@ class DashboardController extends Controller
     {
         $pipeline_id = $request->pipeline;
 
-        $response = Http::withBasicAuth($this->harvey_secret, '')
-            ->timeout($this->timeout)
-            ->get("http://$this->harvey_domain/pipelines/$pipeline_id");
-        $pipeline = $response->successful() ? $response->json() : null;
+        try {
+            $response = Http::withBasicAuth($this->harvey_secret, '')
+                ->timeout($this->timeout)
+                ->get("$this->harvey_domain_protocol://$this->harvey_domain/pipelines/$pipeline_id");
+            $pipeline = $response->successful() ? $response->json() : null;
+        } catch (Throwable $error) {
+            $pipeline = null;
+        }
 
         return view('harvey-pipeline', compact('pipeline'));
     }
@@ -74,15 +98,23 @@ class DashboardController extends Controller
     {
         $project = $request->project;
 
-        $lock_response = Http::withBasicAuth($this->harvey_secret, '')
-            ->timeout($this->timeout)
-            ->get("http://$this->harvey_domain/locks/$project");
-        $locked = $lock_response->successful() ? $lock_response->json()['locked'] : null;
+        try {
+            $lock_response = Http::withBasicAuth($this->harvey_secret, '')
+                ->timeout($this->timeout)
+                ->get("$this->harvey_domain_protocol://$this->harvey_domain/locks/$project");
+            $locked = $lock_response->successful() ? $lock_response->json()['locked'] : null;
+        } catch (Throwable $error) {
+            $locked = null;
+        }
 
-        $project_response = Http::withBasicAuth($this->harvey_secret, '')
-            ->timeout($this->timeout)
-            ->get("http://$this->harvey_domain/pipelines?project=$project");
-        $pipelines = $project_response->successful() ? $project_response->json()['pipelines'] : null;
+        try {
+            $project_response = Http::withBasicAuth($this->harvey_secret, '')
+                ->timeout($this->timeout)
+                ->get("$this->harvey_domain_protocol://$this->harvey_domain/pipelines?project=$project"); // TODO: Add page_size url param here
+            $pipelines = $project_response->successful() ? $project_response->json()['pipelines'] : null;
+        } catch (Throwable $error) {
+            $pipelines = null;
+        }
 
         return view('harvey-project', compact('project', 'locked', 'pipelines'));
     }
@@ -96,9 +128,17 @@ class DashboardController extends Controller
     {
         $project = $request->project;
 
-        $locks_response = Http::withBasicAuth($this->harvey_secret, '')
-            ->timeout($this->timeout)
-            ->put("http://$this->harvey_domain/projects/$project/unlock");
+        try {
+            $locks_response = Http::withBasicAuth($this->harvey_secret, '')
+                ->timeout($this->timeout)
+                ->put("$this->harvey_domain_protocol://$this->harvey_domain/projects/$project/unlock");
+
+            $flash_type = $locks_response->successful() ? 'message' : 'error';
+            $flash_message = $locks_response->successful() ? 'Project unlocked successfully!' : json_encode($locks_response->json());
+        } catch (Throwable $error) {
+            $flash_type = 'error';
+            $flash_message = "Sorry, there was a problem unlocking the project: $error";
+        }
 
         $flash_type = $locks_response->successful() ? 'message' : 'error';
         $flash_message = $locks_response->successful() ? 'Project unlocked successfully!' : json_encode($locks_response->json());
@@ -116,12 +156,17 @@ class DashboardController extends Controller
     {
         $project = $request->project;
 
-        $locks_response = Http::withBasicAuth($this->harvey_secret, '')
-            ->timeout($this->timeout)
-            ->put("http://$this->harvey_domain/projects/$project/lock");
+        try {
+            $locks_response = Http::withBasicAuth($this->harvey_secret, '')
+                ->timeout($this->timeout)
+                ->put("$this->harvey_domain_protocol://$this->harvey_domain/projects/$project/lock");
 
-        $flash_type = $locks_response->successful() ? 'message' : 'error';
-        $flash_message = $locks_response->successful() ? 'Project locked successfully!' : json_encode($locks_response->json());
+            $flash_type = $locks_response->successful() ? 'message' : 'error';
+            $flash_message = $locks_response->successful() ? 'Project locked successfully!' : json_encode($locks_response->json());
+        } catch (Throwable $error) {
+            $flash_type = 'error';
+            $flash_message = "Sorry, there was a problem locking the project: $error";
+        }
 
         session()->flash($flash_type, $flash_message);
         return redirect()->back();