+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Adds authentication to Harvey API with basic auth when present
 - Adds `lock` and `unlock` functionality for projects
+- Wraps all HTTP requests in try/catch blocks to gracefully handle exceptions
 - Adds reusable flash messaging
 - Bumps default timeout from 5 to 10 seconds